app/requirements.txt
xlwt
django
pymysql
django-environ
django-rest-framework
django-ckeditor
django-crontab
django-db-logger
drf-yasg
django-cleanup
pytz
django-cors-headers
cron-descriptor
pyfcm
pandas
weasyprint==52.2